Fix signature of rav::byte_stream::write (RAV-36)

+++ src/ravennakit/core/streams/byte_stream.cpp
@@ -47,8 +47,7 @@ bool rav::byte_stream::exhausted() const {
     return read_position_ >= data_.size();
 }
 
-tl::expected<unsigned long, rav::output_stream::error>
-rav::byte_stream::write(const uint8_t* buffer, const size_t size) {
+tl::expected<size_t, rav::output_stream::error> rav::byte_stream::write(const uint8_t* buffer, const size_t size) {
     try {
         if (write_position_ + size > data_.size()) {
             data_.resize(write_position_ + size, 0);
